Update calculadorabhaskara.py
esqueci de mudar, só aceita numeros inteiros

+++ calculadorabhaskara.py
@@ -20,9 +20,9 @@ def formatar(z):
 
 while True:
     print((' ' * 3), 'Calculadora 2° grau')
-    a = int(input("Digite o valor de a: "))
-    b = int(input("Digite o valor de b: "))
-    c = int(input("Digite o valor de c: "))
+    a = float(input("Digite o valor de a: "))
+    b = float(input("Digite o valor de b: "))
+    c = float(input("Digite o valor de c: "))
     if a == 0:
         print("Isso não é uma equação do segundo grau!")
         continue
